(incomplete) fixed all errors

+++ src/Phantonia.Historia.Language/SyntaxAnalysis/OutcomeDeclarationInfo.cs
@@ -3,4 +3,4 @@ using System.Collections.Immutable;
 
 namespace Phantonia.Historia.Language.SyntaxAnalysis;
 
-public readonly record struct OutcomeDeclarationInfo(Token OutcomeKeyword, Token Name, Token OpenParenthesis, ImmutableArray<Token> Options, ImmutableArray<Token> Commas, Token ClosedParenthesis, Token? DefaultKeyword, Token? DefaultOption, Token Semicolon, int Index);
+public readonly record struct OutcomeDeclarationInfo(Token OutcomeKeyword, Token Name, Token OpenParenthesis, ImmutableArray<Token> Options, ImmutableArray<Token> Commas, Token ClosedParenthesis, Token? DefaultKeyword, Token? DefaultOption, Token Semicolon, long Index);
